+++ src/common/web3/get/queryPriceResolvedEvents.ts
