+++ src/common/web3/get/queryRewardsRetrievedEvents.ts
