+++ src/features/vote/RevealPhase.tsx
@@ -3,15 +3,15 @@ import tw from "twin.macro"; // eslint-disable-line
 import web3 from "web3";
 import { Wrapper } from "./styled/RevealPhase.styled";
 import Button from "common/components/button";
-import { EncryptedVote } from "web3/get/queryEncryptedVotesEvents";
-import { PendingRequest } from "web3/get/queryGetPendingRequests";
-import { revealVotes } from "web3/post/revealVotes";
+import { EncryptedVote } from "common/web3/get/queryEncryptedVotesEvents";
+import { PendingRequest } from "common/web3/get/queryGetPendingRequests";
+import { revealVotes } from "common/web3/post/revealVotes";
 import { useVotingContract } from "hooks";
 import { OnboardContext } from "common/context/OnboardContext";
-import { Round } from "web3/get/queryRounds";
+import { Round } from "common/web3/get/queryRounds";
 import { getMessageSignatureMetamask } from "common/tempUmaFunctions";
-import { snapshotCurrentRound } from "web3/post/snapshotCurrentRound";
-import { VoteRevealed } from "web3/get/queryVotesRevealedEvents";
+import { snapshotCurrentRound } from "common/web3/post/snapshotCurrentRound";
+import { VoteRevealed } from "common/web3/get/queryVotesRevealedEvents";
 import { ModalState } from "./ActiveRequests";
 import useTableValues from "./useTableValues";
 import { ErrorContext } from "common/context/ErrorContext";
